Changed structure of controller to avoid more requires than necessary
Alright, so this is a pretty big change to the structure of the game. The reason I did this was to avoid importing agents like Console_model over and over again. This is so across the game, we deal with only one instance of these agents in order to have consistent data shared across themselves with the use of the controller